[v6.3.0] 코드 정리
