+++ src/decorator.tsx
@@ -1,6 +1,9 @@
 import { useEffect, makeDecorator } from "@storybook/addons";
+
 import setUserAgent from "./core";
+
 import { getUserAgent } from "./utils";
+
 import { DECORATOR_TITLE, PARAM_KEY } from "./constants";
 
 export default makeDecorator({
@@ -8,9 +11,7 @@ export default makeDecorator({
   parameterName: PARAM_KEY,
   wrapper: (storyFn, context) => {
     const userAgent = getUserAgent(context.args);
-
     useEffect(() => setUserAgent(userAgent), [userAgent]);
-
     return storyFn(context);
   },
 });
