+++ src/preset/manager.ts
@@ -1,13 +1,8 @@
 import { addons, types } from "@storybook/addons";
 
-import { Tool } from "../Tool";
+import { Tool } from "../components/Tool";
 
-import {
-  TOOL_ID,
-  ADDON_ID,
-  PARAM_KEY,
-  TOOL_TITLE,
-} from "../constants";
+import { TOOL_ID, ADDON_ID, PARAM_KEY, TOOL_TITLE } from "../constants";
 
 addons.register(ADDON_ID, () => {
   addons.add(TOOL_ID, {
